requirements.txt
diff-match-patch==20121119
Django==1.7.7
django-import-export==0.2.6
tablib==0.10.0
psycopg2==2.5.4
